+++ src/implementations/utils.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
 
-pub fn random_array<const SIZE : usize>(value_min : usize, value_max : usize) -> [usize; SIZE] {
+pub fn random_array<const SIZE : usize>(value_min : i32, value_max : i32) -> [i32; SIZE] {
     let mut rng = rand::thread_rng();
     let mut array = [0; SIZE];
 
@@ -10,6 +10,6 @@ pub fn random_array<const SIZE : usize>(value_min : usize, value_max : usize) ->
     array
 }
 
-pub fn dot_product<const COUNT: usize>(array0 : [usize; COUNT], array1 : [usize; COUNT]) -> usize {
+pub fn dot_product<const COUNT: usize>(array0 : [i32; COUNT], array1 : [i32; COUNT]) -> i32 {
     array0.iter().zip(array1.iter()).map(|(x, y)| x * y).sum()
 }
